feat: add return types to functions
